+++ src/app/(dashboard)/(admin)/users/components/list/filters/filters.tsx
@@ -1,0 +1,11 @@
+import ListFilters from '@/components/list/filters/filters';
+
+import UsersListFiltersRole from './role';
+
+export default function UsersListFilters() {
+  return (
+    <ListFilters>
+      <UsersListFiltersRole />
+    </ListFilters>
+  );
+}